Add the option "overlapEigenvaluesThreshold" which is equivalent to the GAMESS option QMTTOL
QMTTOL : linear dependence threshhold
Any functions in the SALC variational space whose eigenvalue of the overlap matrix is below this tolerence is considered to be linearly dependent. Such functions are dropped from the variational space. What is dropped is not individual basis functions, but rather some linear combination(s) of the entire basis set that represent the linear dependent part of the function space. The default is a reasonable value for most purposes, 1.0E-6.