Fixed overlap eigen issue (#126). Added the input that generated the error as a test of QDO calculations with direct integrals. Added a routine to compute the RMS radius of the quantum particles. Modified the harmonic integrals, shifting the primivites such that the potential is always at the origin, to reduce errors with large exponents
Updated units (#124), replaced AMSTRONG with ANGSTROM in the code, Added eV as a unit.

The change of units slightly modified the results of some tests, for now I increased the thresholds. I plan to review the test reproducilibity in the near future, to make the thresholds less arbitrary

Created a common python file for the functions that are regularly used in the test set