Fixed the growing approach in SCI.
I blame the papers for not explaining this clearly... Basically, the target space is generated by merging the amplitudes and the fixed core space, which are refined by the EN equation and the exact diagonalization approaches, respectively.

Now the eigenvectors convergence is more stable... and now the guess eigenvector are consistent between iterations.

I will clean the code in the next commit, the test are probably failing because the SCI and PT2 values changed, the sum should be similar.